+++ src/backend/playback/base_engine.py
@@ -3,11 +3,12 @@ from abc import ABC, abstractmethod
 from src.sentinels import Sentinel
 from src.config import CONFIG
 
-class Engine(ABC):
+class BaseEngine(ABC):
     def __init__(self, logger, name, on_end_func):
         self.logger = logger
         self.name = name
         self.on_end = on_end_func
+        self.medias: list = []
         self.number: int = 0
         self.volume: int = CONFIG.default_volume
         self.mute: bool = False
@@ -17,23 +18,11 @@ class Engine(ABC):
         ...
 
     @abstractmethod
-    def get_media_len(self):
+    def get_progress(self) -> tuple[int, int]: # (time, length)
         ...
 
     @abstractmethod
-    def get_progress(self) -> dict[str, int]:
-        ...
-
-    @abstractmethod
-    def switch_prev(self) -> Sentinel:
-        ...
-
-    @abstractmethod
-    def switch_next(self) -> Sentinel:
-        ...
-
-    @abstractmethod
-    def load_number(self, num) -> Sentinel:
+    def load_number(self) -> Sentinel:
         ...
 
     @abstractmethod
@@ -53,10 +42,6 @@ class Engine(ABC):
         ...
 
     @abstractmethod
-    def toggle(self) -> Sentinel:
-        ...
-
-    @abstractmethod
     def pause(self) -> Sentinel:
         ...
 
@@ -65,11 +50,7 @@ class Engine(ABC):
         ...
 
     @abstractmethod
-    def set_volume(self, volume) -> Sentinel:
-        ...
-
-    @abstractmethod
-    def set_mute(self, mute) -> Sentinel:
+    def apply_volume(self) -> Sentinel:
         ...
     
     @abstractmethod