+++ src/backend/playback/vlc_engine.py
@@ -21,9 +21,9 @@ from src.constants import PLAYER_POLL_INTERVAL
 from src.config import CONFIG
 from src.sentinels import SENTINELS
 from src.error import InitializationError
-from .engine import Engine
+from .base_engine import BaseEngine
 
-class VLCEngine(Engine):
+class VLCEngine(BaseEngine):
     def __init__(self, logger, on_end_func):
         super().__init__(logger, 'VLC', on_end_func)
         if vlc is None:
@@ -31,7 +31,6 @@ class VLCEngine(Engine):
         else:
             self.instance = vlc.Instance('--no-video')
             self.player = self.instance.media_player_new()
-            self.medias = []
             self._attach_events()
             self.logger.debug(f'{__name__} initiated')
 
@@ -57,58 +56,28 @@ class VLCEngine(Engine):
             vlc.State.Stopped: SENTINELS.STOPPED
             }.get(self.player.get_state(), SENTINELS.PLAYER_INVALID)
 
-    def get_media_len(self):
-        return len(self.medias)
-
     def get_progress(self):
-        length = self.player.get_length()
         time_ = self.player.get_time()
-        return {'length': length, 'time': time_}
+        length = self.player.get_length()
+        return time_, length
 
-    def switch_prev(self):
-        return self.load_number(self.number - 1)
-
-    def switch_next(self):
-        return self.load_number(self.number + 1)
-
-    def load_number(self, num):
-        if len(self.medias) > 0:
-            if num < 0:
-                self.number = len(self.medias) - 1
-            elif num >= len(self.medias):
-                self.number = 0
-            else:
-                self.number = num
-
-            self.player.set_media(self.medias[self.number])
-            return self.play()
-        else:
-            return SENTINELS.PLAYER_EMPTY
+    def load_number(self):
+        self.player.set_media(self.medias[self.number])
+        return SENTINELS.SUCCESS
 
     def load_paths(self, paths):
-        if not isinstance(paths, (list, tuple)):
-            paths = [paths]
-
-        if len(paths) == 0:
-            self.logger.error('Can not load empty path list')
-            return SENTINELS.PLAYER_LOAD_EMPTY
-        else:
-            self.medias = []
-            for path in paths:
-                media = self.instance.media_new(path)
-                media.parse()
-                if media.get_state() == vlc.State.Error or media.get_parsed_status() == vlc.MediaParsedStatus.failed:
-                    self.logger.error(f'Failed to parse {path}')
-                    break
-                else:
-                    self.medias.append(media)
+        self.medias = []
+        for path in paths:
+            media = self.instance.media_new(path)
+            media.parse()
+            if media.get_state() == vlc.State.Error or media.get_parsed_status() == vlc.MediaParsedStatus.failed:
+                self.logger.error(f'Failed to parse {path}')
+                self.medias = []
+                return SENTINELS.ENGINE_ERROR
             else:
-                self.logger.info(f'Opened {len(paths)} files: {", ".join(paths)}')
-                self.number = 0
-                return self.load_number(0)                 
+                self.medias.append(media)
 
-            self.medias = []
-            return SENTINELS.ENGINE_ERROR
+        return SENTINELS.SUCCESS
 
     def stop(self):
         self.player.stop()
@@ -125,16 +94,8 @@ class VLCEngine(Engine):
             return SENTINELS.ENGINE_ERROR
 
     def jump_pos(self, pos):
-        if self.player.get_state() in (vlc.State.Playing, vlc.State.Paused):
-            length = self.get_progress()['length']
-            pos = int(pos)
-            if pos > length:
-                return SENTINELS.POS_TOO_LATE
-            else:
-                self.player.set_time(pos)
-                return SENTINELS.SUCCESS
-        else:
-            return SENTINELS.INVALID_PLAYER_STATE
+        self.player.set_time(pos)
+        return SENTINELS.SUCCESS
 
     def play(self):
         self.player.play()
@@ -144,77 +105,48 @@ class VLCEngine(Engine):
 
         elif result == vlc.State.Playing:
             self.logger.info('Started playing')
-            self._apply_volume()
+            self.apply_volume()
             return SENTINELS.SUCCESS
         
         elif result == vlc.State.Error:
             self.logger.error('Failed to start playing')
             return SENTINELS.ENGINE_ERROR
 
-    def toggle(self):
-        state = self.player.get_state()
-        if state == vlc.State.Paused:
-            return self.resume()
-        elif state == vlc.State.Playing:
-            return self.pause()
-        else:
-            self.logger.error('Invalid player state, can not toggle')
-            return SENTINELS.INVALID_PLAYER_STATE
-
     def pause(self):
-        if self.player.get_state() == vlc.State.Playing:
-            self.player.set_pause(1)
-            result = self._wait_state([vlc.State.Paused])
-            if result is None:
-                return SENTINELS.PLAYER_TIMEOUT
+        self.player.set_pause(1)
+        result = self._wait_state([vlc.State.Paused])
+        if result is None:
+            return SENTINELS.PLAYER_TIMEOUT
 
-            elif result == vlc.State.Paused:
-                self.logger.info('Paused')
-                return SENTINELS.SUCCESS
+        elif result == vlc.State.Paused:
+            self.logger.info('Paused')
+            return SENTINELS.SUCCESS
 
-            elif result == vlc.State.Error:
-                self.logger.error('Failed to pause audio')
-                return SENTINELS.ENGINE_ERROR
-        else:
-            self.logger.error('Failed to pause because player is not playing')
-            return SENTINELS.INVALID_PLAYER_STATE
+        elif result == vlc.State.Error:
+            self.logger.error('Failed to pause audio')
+            return SENTINELS.ENGINE_ERROR
 
     def resume(self):
-        if self.player.get_state() == vlc.State.Paused:
-            self.player.set_pause(0)
-            result = self._wait_state([vlc.State.Playing])
-            if result is None:
-                return SENTINELS.PLAYER_TIMEOUT
-            elif result == vlc.State.Playing:
-                self.logger.info('Resumed')
-                return SENTINELS.SUCCESS
+        self.player.set_pause(0)
+        result = self._wait_state([vlc.State.Playing])
+        if result is None:
+            return SENTINELS.PLAYER_TIMEOUT
+        elif result == vlc.State.Playing:
+            self.logger.info('Resumed')
+            return SENTINELS.SUCCESS
 
-            elif result == vlc.State.Error:
-                self.logger.info('Failed to resume audio')
-                return SENTINELS.ENGINE_ERROR
-        else:
-            self.logger.error('Failed to resume because player is not paused')
-            return SENTINELS.INVALID_PLAYER_STATE
-
-    def set_volume(self, volume):
-        self.volume = volume
-        self.logger.debug(f'Set target volume to {self.volume}')
-        self._apply_volume()
-        return SENTINELS.SUCCESS
+        elif result == vlc.State.Error:
+            self.logger.info('Failed to resume audio')
+            return SENTINELS.ENGINE_ERROR
     
-    def set_mute(self, mute):
-        self.mute = mute
-        self.logger.debug(f'Set target mute mode to {self.mute}')
-        self._apply_volume()
-        return SENTINELS.SUCCESS
-    
-    def _apply_volume(self):
+    def apply_volume(self):
         if self.player.get_state() in (vlc.State.Playing, vlc.State.Paused):
             self.logger.debug(f'Applied volume: {self.volume}, mute: {self.mute}')
             self.player.audio_set_volume(self.volume)
             self.player.audio_set_mute(self.mute)
         else:
             self.logger.debug(f'Can not apply volume and mute: unsupported player state')
+        return SENTINELS.SUCCESS
         
     def on_exit(self):
         self.player.stop()